chaging telephone number to type bigint

+++ src/main/java/up/edu/microservicios/dao/PacienteDAOH2.java
@@ -30,7 +30,7 @@ public class PacienteDAOH2 implements iDao<Paciente>{
 
             ps.setString(1, paciente.getNombre());
             ps.setString(2, paciente.getApellido());
-            ps.setInt(3, paciente.getNumeroContacto());
+            ps.setLong(3, paciente.getNumeroContacto());
             ps.setDate(4, Date.valueOf(paciente.getFechaIngreso())); // si tu atributo es LocalDate
             ps.setInt(5, paciente.getDomicilio().getId());
             ps.setString(6, paciente.getEmail());
@@ -66,7 +66,7 @@ public class PacienteDAOH2 implements iDao<Paciente>{
             DomicilioDAOH2 daoAux= new DomicilioDAOH2();
             while(rs.next()){
                 domicilio=daoAux.buscar(rs.getInt(6));
-                paciente= new Paciente(rs.getInt(1),rs.getString(2),rs.getString(3),rs.getInt(4),rs.getDate(5).toLocalDate(),domicilio,rs.getString(7));
+                paciente= new Paciente(rs.getInt(1),rs.getString(2),rs.getString(3),rs.getLong(4),rs.getDate(5).toLocalDate(),domicilio,rs.getString(7));
             }
         }catch (Exception e){
             e.getMessage();
@@ -98,7 +98,7 @@ public class PacienteDAOH2 implements iDao<Paciente>{
             PreparedStatement ps_update = connection.prepareStatement(SQL_UPDATE_BY_ID);
             ps_update.setString(1, paciente.getNombre());
             ps_update.setString(2, paciente.getApellido());
-            ps_update.setInt(3, paciente.getNumeroContacto());
+            ps_update.setLong(3, paciente.getNumeroContacto());
             ps_update.setDate(4, java.sql.Date.valueOf(paciente.getFechaIngreso()));
             ps_update.setInt(5, paciente.getDomicilio().getId());
             ps_update.setString(6, paciente.getEmail());
@@ -127,7 +127,7 @@ public class PacienteDAOH2 implements iDao<Paciente>{
             DomicilioDAOH2 daoAux= new DomicilioDAOH2();
             while(rs.next()){
                 domicilio=daoAux.buscar(rs.getInt(6));
-                Paciente p = new Paciente(rs.getInt(1),rs.getString(2),rs.getString(3),rs.getInt(4),rs.getDate(5).toLocalDate(),domicilio,rs.getString(7));
+                Paciente p = new Paciente(rs.getInt(1),rs.getString(2),rs.getString(3),rs.getLong(4),rs.getDate(5).toLocalDate(),domicilio,rs.getString(7));
                 pacientes.add(p);
             }
         }catch (Exception e){
